+++ src/util/IncognitoBrowser.cpp
@@ -1,88 +1,93 @@
 #include "util/IncognitoBrowser.hpp"
 #ifdef USEWINSDK
 #    include "util/WindowsHelper.hpp"
+#elif defined(Q_OS_UNIX) and !defined(Q_OS_DARWIN)
+#    include "util/XDGHelper.hpp"
 #endif
 
 #include <QProcess>
-#include <QRegularExpression>
 #include <QVariant>
 
 namespace {
 
 using namespace chatterino;
 
-#ifdef USEWINSDK
-QString injectPrivateSwitch(QString command)
+QString getPrivateSwitch(const QString &browserExecutable)
 {
     // list of command line switches to turn on private browsing in browsers
     static auto switches = std::vector<std::pair<QString, QString>>{
-        {"firefox", "-private-window"},  {"librewolf", "-private-window"},
-        {"waterfox", "-private-window"}, {"icecat", "-private-window"},
-        {"chrome", "-incognito"},        {"vivaldi", "-incognito"},
-        {"opera", "-newprivatetab"},     {"opera\\\\launcher", "--private"},
-        {"iexplore", "-private"},        {"msedge", "-inprivate"},
+        {"firefox", "-private-window"},     {"librewolf", "-private-window"},
+        {"waterfox", "-private-window"},    {"icecat", "-private-window"},
+        {"chrome", "-incognito"},           {"vivaldi", "-incognito"},
+        {"opera", "-newprivatetab"},        {"opera\\launcher", "--private"},
+        {"iexplore", "-private"},           {"msedge", "-inprivate"},
+        {"firefox-esr", "-private-window"}, {"chromium", "-incognito"},
     };
 
-    // transform into regex and replacement string
-    std::vector<std::pair<QRegularExpression, QString>> replacers;
+    // compare case-insensitively
+    auto lowercasedBrowserExecutable = browserExecutable.toLower();
+
+#ifdef Q_OS_WINDOWS
+    if (lowercasedBrowserExecutable.endsWith(".exe"))
+    {
+        lowercasedBrowserExecutable.chop(4);
+    }
+#endif
+
     for (const auto &switch_ : switches)
     {
-        replacers.emplace_back(
-            QRegularExpression("(" + switch_.first + "\\.exe\"?).*",
-                               QRegularExpression::CaseInsensitiveOption),
-            "\\1 " + switch_.second);
-    }
-
-    // try to find matching regex and apply it
-    for (const auto &replacement : replacers)
-    {
-        if (replacement.first.match(command).hasMatch())
+        if (lowercasedBrowserExecutable.endsWith(switch_.first))
         {
-            command.replace(replacement.first, replacement.second);
-            return command;
+            return switch_.second;
         }
     }
 
     // couldn't match any browser -> unknown browser
-    return QString();
+    return {};
 }
 
-QString getCommand()
+QString getDefaultBrowserExecutable()
 {
+#ifdef USEWINSDK
     // get default browser start command, by protocol if possible, falling back to extension if not
     QString command =
-        getAssociatedCommand(AssociationQueryType::Protocol, L"http");
+        getAssociatedExecutable(AssociationQueryType::Protocol, L"http");
 
     if (command.isNull())
     {
         // failed to fetch default browser by protocol, try by file extension instead
-        command =
-            getAssociatedCommand(AssociationQueryType::FileExtension, L".html");
+        command = getAssociatedExecutable(AssociationQueryType::FileExtension,
+                                          L".html");
     }
 
     if (command.isNull())
     {
         // also try the equivalent .htm extension
-        command =
-            getAssociatedCommand(AssociationQueryType::FileExtension, L".htm");
-    }
-
-    if (command.isNull())
-    {
-        // failed to find browser command
-        return QString();
-    }
-
-    // inject switch to enable private browsing
-    command = injectPrivateSwitch(command);
-    if (command.isNull())
-    {
-        return QString();
+        command = getAssociatedExecutable(AssociationQueryType::FileExtension,
+                                          L".htm");
     }
 
     return command;
-}
+#elif defined(Q_OS_UNIX) && !defined(Q_OS_DARWIN)
+    static QString defaultBrowser = []() -> QString {
+        auto desktopFile = getDefaultBrowserDesktopFile();
+        if (desktopFile.has_value())
+        {
+            auto entry = desktopFile->getEntries("Desktop Entry");
+            auto exec = entry.find("Exec");
+            if (exec != entry.end())
+            {
+                return parseDesktopExecProgram(exec->second.trimmed());
+            }
+        }
+        return {};
+    }();
+
+    return defaultBrowser;
+#else
+    return {};
 #endif
+}
 
 }  // namespace
 
@@ -90,23 +95,15 @@ namespace chatterino {
 
 bool supportsIncognitoLinks()
 {
-#ifdef USEWINSDK
-    return !getCommand().isNull();
-#else
-    return false;
-#endif
+    auto browserExe = getDefaultBrowserExecutable();
+    return !browserExe.isNull() && !getPrivateSwitch(browserExe).isNull();
 }
 
 bool openLinkIncognito(const QString &link)
 {
-#ifdef USEWINSDK
-    auto command = getCommand();
-
-    // TODO: split command into program path and incognito argument
-    return QProcess::startDetached(command, {link});
-#else
-    return false;
-#endif
+    auto browserExe = getDefaultBrowserExecutable();
+    return QProcess::startDetached(browserExe,
+                                   {getPrivateSwitch(browserExe), link});
 }
 
 }  // namespace chatterino